BMAT_test/requirements.txt
Django>=3.0
django-extensions
django-filter
djangorestframework
ipdb
ipython
markdown
metaphone
psycopg2>=2.8.4
requests
unidecode